requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
discord.py==2.1.0
docopt==0.6.2
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
mysql-connector==2.2.9
ordered-set==4.1.0
pipreqs==0.4.13
PyLaTeX==1.4.1
pyodbc==4.0.35
requests==2.31.0
urllib3==2.0.4
yarg==0.1.9
yarl==1.8.1
mysql-connector-python==8.0.31
